funciones

+++ project/ASTNodes/Number.cs
@@ -1,12 +1,12 @@
 public class Number : ASTNode
 {
     public Token Num { get; private set; }
-    public double Value { get; private set; }
+    public int Value { get; private set; }
 
     public Number(Token num)
     {
         Num = num;
-        Value = double.Parse(num.Lexeme);
+        Value = int.Parse(num.Lexeme);
     }
 
      public override T Accept<T>(IVisitor<T> visitor)
